Update rplidar.py
Receiving start_angle errors when using a Slamtec RPLIDAR A1M8 on Windows/Python and attemping to set scan_mode to 'express'. For some reason (I'm new to this) self.express_data is being set to bool. Added to the "if not self.express_data" check to see if it's bool. Works fine now in express mode.